requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==3.2.24
django-allauth==0.41.0
django-crispy-forms==1.14.0
gunicorn==21.2.0
oauthlib==3.2.2
psycopg2==2.9.9
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.3.1
sqlparse==0.4.4
